+++ ICD.Connect.Conferencing/ConferencePoints/IConferencePointSettings.cs
@@ -1,10 +1,19 @@
-﻿using ICD.Connect.Conferencing.EventArguments;
+﻿using ICD.Connect.Audio.VolumePoints;
+using ICD.Connect.Conferencing.EventArguments;
 using ICD.Connect.Devices.Points;
 
 namespace ICD.Connect.Conferencing.ConferencePoints
 {
 	public interface IConferencePointSettings : IPointSettings
 	{
+		/// <summary>
+		/// The type of call to use the conference control for.
+		/// </summary>
 		eCallType Type { get; set; }
+
+		/// <summary>
+		/// Determines if the privacy mute control will be driven by the control system, and/or drive the control system.
+		/// </summary>
+		ePrivacyMuteFeedback PrivacyMuteMask { get; set; }
 	}
 }